% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimfa.R
\name{install_nimfa}
\alias{install_nimfa}
\title{Load nimfa library in R}
\usage{
install_nimfa(virtualenv = "BiocNimfa")
}
\arguments{
\item{virtualenv}{Name of virtualenv}
}
\value{
The nimfa module as a python builtin module
}
\description{
Load nimfa library in R
}
\examples{
nimfa = install_nimfa()
v <- nimfa$examples$medulloblastoma$read(normalize=TRUE)
lsnmf <- nimfa$Lsnmf(v, seed="random_vcol", rank=50L, max_iter=100L)
lsnmf_fit <- lsnmf()
print(lsnmf_fit$fit$rss())
}
