% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimfa.R
\name{load_nimfa}
\alias{load_nimfa}
\title{Load nimfa library in R}
\usage{
load_nimfa(virtualenv = "BiocNimfa")
}
\description{
Load nimfa library in R
}
\examples{
nimfa = load_nimfa()
v <- nimfa$examples$medulloblastoma$read(normalize=TRUE)
lsnmf <- nimfa$Lsnmf(v, seed="random_vcol", rank=50L, max_iter=100L)
lsnmf_fit <- lsnmf()
print(lsnmf_fit$fit$rss())
}
