% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimfa.R
\name{install_nimfa}
\alias{install_nimfa}
\alias{nimfa}
\title{Import python nimfa library for use in R}
\usage{
install_nimfa(envname = "BiocNimfa")

nimfa(envname = "BiocNimfa")
}
\arguments{
\item{envname}{character(1) name of the virtual environment
(usually under `~/.virtualenvs` where Nimfa and its python
dependencies are available.}
}
\value{
The nimfa module as a python builtin module
}
\description{
Import python nimfa library for use in R
}
\examples{
nimfa = import_nimfa()
v <- nimfa$examples$medulloblastoma$read(normalize=TRUE)
lsnmf <- nimfa$Lsnmf(v, seed="random_vcol", rank=50L, max_iter=100L)
lsnmf_fit <- lsnmf()
lsnmf_fit$fit$rss()
}
